    \name{data_list}
     \docType{data}
     \alias{data_list}
     \title{Simulated MCMC sample and related information}
     \description{This is a (very) small MCMC sample corresponding to data of 5 observations from a mixture 2 normal distributions. The MCMC sample consists of 300 iterations. It is stored to \code{data_list$mcmc.pars}. \code{data_list$mcmc.pars[,,1]} corresponds to means, \code{data_list$mcmc.pars[,,2]} corresponds to variances and \code{data_list$mcmc.pars[,,3]} corresponds to weights.}
     \usage{data_list}
     \format{A list containing simulated MCMC sample and all information required for the relabelling algorithms.}
     \keyword{datasets}
