    \name{lamb}
     \docType{data}
     \alias{lamb}
     \title{Fetal lamb dataset}
     \description{240 body movement measurements of a fetal lamb at consecutive 5 second intervals.}
     \usage{lamb}
     \format{Count data.}
     \keyword{datasets}
     \references{
           %% ~put references to the literature/web site here ~
Leroux B, Putterman M (1992). Maximum Penalized Likelihood estimation for independent and Markov-dependent Mixture models.Biometrics, 48, 545--558.
}

